refactor: extract getTodayDate function to utils/dateUtils.ts
- Move getTodayDate function from ExpenseCreateModal.tsx to utils/dateUtils.ts
- Make the function reusable across the codebase
- Add JSDoc documentation for the function
